pyproject.toml
[tool.poetry]
name = "nikaritds"
version = "0.1.0"
description = ""
authors = ["Daniel Alfaro <daniel.alfaro@ignis.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
firebase-admin = "^6.6.0"
pandas = "^2.2.3"
notebook = "^7.3.2"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
jupyter-core = "^5.7.2"
pandas-gbq = "^0.26.0"
pyarrow = "^18.1.0"
google-cloud = "^0.34.0"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
